scripts/requirements.txt
# Face Matching DNI - Dependencias Python
face_recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
pyzbar==0.1.9
dlib==19.24.2
